Create dfp_300x250_ros_ad
This is the Document Header code and Document Body Code for the DFP ad Server. Copy and paste the first tag in to the <head> section of the HTML. The second is the ad unit tag. copy and paste this in to the <body> section of the HTML. Place this tag in the desired location of the 300x250 ad unit